+++ src/js/ui/NewDataExploration/AdvancedLayers.js
@@ -2,7 +2,7 @@ import React, {useState} from 'react';
 import {Typography} from "@material-ui/core";
 import Box from "@material-ui/core/Box";
 
-export default function AdvancedMenu() {
+export default function AdvancedLayers() {
     return (
         <Box>
             <Typography>
