+++ src/js/ui/NewDataExploration/LayerCheckbox.js
@@ -3,7 +3,7 @@ import FormGroup from '@material-ui/core/FormGroup';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import Checkbox from '@material-ui/core/Checkbox';
 
-export default function DECheckbox(props) {
+export default function LayerCheckbox(props) {
     const [state, setState] = React.useState({
         checkedMe: true,
     });
