+++ src/js/ui/NewDataExploration/LayerSlider.js
@@ -18,7 +18,7 @@ const useStyles = makeStyles({
 
 //FIXME include a conditional render for a single-point slider
 
-export default function DESlider(props) {
+export default function LayerSlider(props) {
     const classes = useStyles();
     const min = props.constraint["range"][0];
     const max = props.constraint["range"][1];
